introduce utils and docs module
