+++ git_machete/cmd.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from typing import Any, Callable, Dict, Generator, Iterable, Iterator, List, Match, Optional, Set, Tuple, TypeVar
+from typing import Any, Callable, Dict, Generator, Iterator, List, Match, Optional, Set, Tuple, TypeVar
 
 from git_machete import __version__
 import datetime
@@ -14,6 +14,9 @@ import subprocess
 import sys
 import textwrap
 
+from git_machete import utils
+from git_machete.docs import short_docs, long_docs
+
 
 # Core utils
 
@@ -26,33 +29,6 @@ class MacheteException(Exception):
 
     def __str__(self) -> str:
         return str(self.parameter)
-
-
-def excluding(iterable: Iterable[T], s: Iterable[T]) -> List[T]:
-    return list(filter(lambda x: x not in s, iterable))
-
-
-def flat_map(func: Callable[[T], List[T]], iterable: Iterable[T]) -> List[T]:
-    return sum(map(func, iterable), [])
-
-
-def map_truthy_only(func: Callable[[T], Optional[T]], iterable: Iterable[T]) -> List[T]:
-    return list(filter(None, map(func, iterable)))
-
-
-def non_empty_lines(s: str) -> List[str]:
-    return list(filter(None, s.split("\n")))
-
-
-# Converts a lambda accepting N arguments to a lambda accepting one argument, an N-element tuple.
-# Name matching Scala's `tupled` on `FunctionX`.
-def tupled(f: Callable[..., T]) -> Callable[[Any], T]:
-    return lambda tple: f(*tple)
-
-
-def get_second(pair: Tuple[str, str]) -> str:
-    a, b = pair
-    return b
 
 
 ENDC = '\033[0m'
@@ -178,7 +154,7 @@ def pretty_choices(*choices: str) -> str:
             return colored(c, RED)
         else:
             return colored(c, ORANGE)
-    return f" ({', '.join(map_truthy_only(format_choice, choices))}) "
+    return f" ({', '.join(utils.map_truthy_only(format_choice, choices))}) "
 
 
 def pick(choices: List[str], name: str, apply_fmt: bool = True) -> str:
@@ -212,27 +188,10 @@ def warn(msg: str, apply_fmt: bool = True) -> None:
         displayed_warnings.add(msg)
 
 
-def directory_exists(path: str) -> bool:
-    try:
-        # Note that os.path.isdir itself (without os.path.abspath) isn't reliable
-        # since it returns a false positive (True) for the current directory when if it doesn't exist
-        return os.path.isdir(os.path.abspath(path))
-    except OSError:
-        return False
-
-
-def current_directory_or_none() -> Optional[str]:
-    try:
-        return os.getcwd()
-    except OSError:
-        # This happens when current directory does not exist (typically: has been deleted)
-        return None
-
-
 # Let's keep the flag to avoid checking for current directory's existence
 # every time any command is being popened or run.
 current_directory_confirmed_to_exist: bool = False
-initial_current_directory: Optional[str] = current_directory_or_none() or os.getenv('PWD')
+initial_current_directory: Optional[str] = utils.current_directory_or_none() or os.getenv('PWD')
 
 
 def mark_current_directory_as_possibly_non_existent() -> None:
@@ -243,13 +202,13 @@ def mark_current_directory_as_possibly_non_existent() -> None:
 def chdir_upwards_until_current_directory_exists(cli_ctxt: CommandLineContext) -> None:
     global current_directory_confirmed_to_exist
     if not current_directory_confirmed_to_exist:
-        current_directory: Optional[str] = current_directory_or_none()
+        current_directory: Optional[str] = utils.current_directory_or_none()
         if not current_directory:
             while not current_directory:
                 # Note: 'os.chdir' only affects the current process and its subprocesses;
                 # it doesn't propagate to the parent process (which is typically a shell).
                 os.chdir(os.path.pardir)
-                current_directory = current_directory_or_none()
+                current_directory = utils.current_directory_or_none()
             debug(cli_ctxt,
                   "chdir_upwards_until_current_directory_exists()",
                   f"current directory did not exist, chdired up into {current_directory}")
@@ -444,7 +403,7 @@ class MacheteClient:
                          pretty_choices("y", "e[dit]", "N"), opt_yes_msg=None)
 
         def recursive_slide_out_invalid_branches(b: str) -> List[str]:
-            new_down_branches = flat_map(recursive_slide_out_invalid_branches, self.down_branches.get(b, []))
+            new_down_branches = utils.flat_map(recursive_slide_out_invalid_branches, self.down_branches.get(b, []))
             if b in invalid_branches:
                 if b in self.down_branches:
                     del self.down_branches[b]
@@ -462,8 +421,8 @@ class MacheteClient:
                 self.down_branches[b] = new_down_branches
                 return [b]
 
-        self.roots = flat_map(recursive_slide_out_invalid_branches, self.roots)
-        self.managed_branches = excluding(self.managed_branches, invalid_branches)
+        self.roots = utils.flat_map(recursive_slide_out_invalid_branches, self.roots)
+        self.managed_branches = utils.excluding(self.managed_branches, invalid_branches)
         if ans in ('y', 'yes'):
             self.save_definition_file()
         elif ans in ('e', 'edit'):
@@ -612,14 +571,14 @@ class MacheteClient:
                 root_of[b] = get_root_of(root_of[b])
             return root_of[b]
 
-        non_root_fixed_branches = excluding(all_local_branches, self.roots)
+        non_root_fixed_branches = utils.excluding(all_local_branches, self.roots)
         last_checkout_timestamps = get_latest_checkout_timestamps(self.cli_ctxt)
         non_root_fixed_branches_by_last_checkout_timestamps = sorted(
             (last_checkout_timestamps.get(b, 0), b) for b in non_root_fixed_branches)
         if self.cli_ctxt.opt_checked_out_since:
             threshold = parse_git_timespec_to_unix_timestamp(self.cli_ctxt, self.cli_ctxt.opt_checked_out_since)
             stale_non_root_fixed_branches = [b for (timestamp, b) in itertools.takewhile(
-                tupled(lambda timestamp, b: timestamp < threshold),
+                utils.tupled(lambda timestamp, b: timestamp < threshold),
                 non_root_fixed_branches_by_last_checkout_timestamps
             )]
         else:
@@ -632,13 +591,13 @@ class MacheteClient:
                      f"only branches checked out at or after ca. <b>{threshold_date}</b> are included.\n"
                      "Use `git machete discover --checked-out-since=<date>` (where <date> can be e.g. `'2 weeks ago'` or `2020-06-01`) "
                      "to change this threshold so that less or more branches are included.\n")
-        self.managed_branches = excluding(all_local_branches, stale_non_root_fixed_branches)
+        self.managed_branches = utils.excluding(all_local_branches, stale_non_root_fixed_branches)
         if self.cli_ctxt.opt_checked_out_since and not self.managed_branches:
             warn(
                 "no branches satisfying the criteria. Try moving the value of `--checked-out-since` further to the past.")
             return
 
-        for b in excluding(non_root_fixed_branches, stale_non_root_fixed_branches):
+        for b in utils.excluding(non_root_fixed_branches, stale_non_root_fixed_branches):
             u = self.infer_upstream(b, condition=lambda candidate: get_root_of(candidate) != b and candidate not in stale_non_root_fixed_branches, reject_reason_message="choosing this candidate would form a cycle in the resulting graph or the candidate is a stale branch")
             if u:
                 debug(self.cli_ctxt, "discover_tree()",
@@ -667,10 +626,10 @@ class MacheteClient:
             warn("skipping %s since %s merged to another branch and would not have any downstream branches.\n"
                  % (", ".join(f"`{b}`" for b in merged_branches_to_skip),
                     "it's" if len(merged_branches_to_skip) == 1 else "they're"))
-            self.managed_branches = excluding(self.managed_branches, merged_branches_to_skip)
+            self.managed_branches = utils.excluding(self.managed_branches, merged_branches_to_skip)
             for b in merged_branches_to_skip:
                 u = self.up_branch[b]
-                self.down_branches[u] = excluding(self.down_branches[u], [b])
+                self.down_branches[u] = utils.excluding(self.down_branches[u], [b])
                 del self.up_branch[b]
             # We're NOT applying the removal process recursively,
             # so it's theoretically possible that some merged branches became childless
@@ -791,7 +750,7 @@ class MacheteClient:
             dds = self.down_branches.get(d, [])
             for dd in dds:
                 self.up_branch[dd] = b
-            self.down_branches[b] = flat_map(
+            self.down_branches[b] = utils.flat_map(
                 lambda bd: dds if bd == d else [bd],
                 self.down_branches[b])
             self.save_definition_file()
@@ -886,7 +845,7 @@ class MacheteClient:
                             nearest_remaining_branch = u
                     for d in self.down_branches.get(b) or []:
                         self.up_branch[d] = u
-                    self.down_branches[u] = flat_map(
+                    self.down_branches[u] = utils.flat_map(
                         lambda ud: (self.down_branches.get(b) or []) if ud == b else [ud],
                         self.down_branches[u])
                     if b in self.annotations:
@@ -1218,10 +1177,10 @@ class MacheteClient:
             warn(f"{first_part}.\n\n{second_part}.")
 
     def delete_unmanaged(self) -> None:
-        branches_to_delete = excluding(local_branches(self.cli_ctxt), self.managed_branches)
+        branches_to_delete = utils.excluding(local_branches(self.cli_ctxt), self.managed_branches)
         cb = current_branch_or_none(self.cli_ctxt)
         if cb and cb in branches_to_delete:
-            branches_to_delete = excluding(branches_to_delete, [cb])
+            branches_to_delete = utils.excluding(branches_to_delete, [cb])
             print(fmt(f"Skipping current branch `{cb}`"))
         if branches_to_delete:
             branches_merged_to_head = merged_local_branches(self.cli_ctxt)
@@ -1300,7 +1259,7 @@ class MacheteClient:
                   "fork_point_and_containing_branch_defs({b})",
                   f"commit {fp_sha} is the most recent point in history of {b} to occur on "
                   "filtered reflog of any other branch or its remote counterpart "
-                  f"(specifically: {' and '.join(map(get_second, containing_branch_defs))})")
+                  f"(specifically: {' and '.join(map(utils.get_second, containing_branch_defs))})")
 
             if u and is_ancestor_or_equal(self.cli_ctxt, u, b) and not is_ancestor_or_equal(self.cli_ctxt, u, fp_sha,
                                                                                             later_prefix=""):
@@ -1636,7 +1595,7 @@ class MacheteClient:
                     def branch_def_to_str(lb: str, lb_or_rb: str) -> str:
                         return lb if lb == lb_or_rb else f"{lb_or_rb} (remote counterpart of {lb})"
 
-                    joined_branch_defs = ", ".join(map(tupled(branch_def_to_str), branch_defs))
+                    joined_branch_defs = ", ".join(map(utils.tupled(branch_def_to_str), branch_defs))
                     yield dim(f"{sha_} => {joined_branch_defs}")
 
             debug(self.cli_ctxt, f"match_log_to_filtered_reflogs({b})",
@@ -1651,16 +1610,16 @@ class MacheteClient:
                 def lb_is_not_b(lb: str, lb_or_rb: str) -> bool:
                     return lb != b
 
-                containing_branch_defs = sorted(filter(tupled(lb_is_not_b), branch_defs), key=get_second)
+                containing_branch_defs = sorted(filter(utils.tupled(lb_is_not_b), branch_defs), key=utils.get_second)
                 if containing_branch_defs:
                     debug(self.cli_ctxt,
                           f"match_log_to_filtered_reflogs({b})",
-                          f"commit {sha} found in filtered reflog of {' and '.join(map(get_second, branch_defs))}")
+                          f"commit {sha} found in filtered reflog of {' and '.join(map(utils.get_second, branch_defs))}")
                     yield sha, containing_branch_defs
                 else:
                     debug(self.cli_ctxt,
                           f"match_log_to_filtered_reflogs({b})",
-                          f"commit {sha} found only in filtered reflog of {' and '.join(map(get_second, branch_defs))}; ignoring")
+                          f"commit {sha} found only in filtered reflog of {' and '.join(map(utils.get_second, branch_defs))}; ignoring")
             else:
                 debug(self.cli_ctxt, f"match_log_to_filtered_reflogs({b})", f"commit {sha} not found in any filtered reflog")
 
@@ -1668,7 +1627,7 @@ class MacheteClient:
         for sha, containing_branch_defs in self.match_log_to_filtered_reflogs(b):
             debug(self.cli_ctxt,
                   f"infer_upstream({b})",
-                  f"commit {sha} found in filtered reflog of {' and '.join(map(get_second, containing_branch_defs))}")
+                  f"commit {sha} found in filtered reflog of {' and '.join(map(utils.get_second, containing_branch_defs))}")
 
             for candidate, original_matched_branch in containing_branch_defs:
                 if candidate != original_matched_branch:
@@ -1785,27 +1744,7 @@ def allowed_directions(allow_current: bool) -> str:
 
 
 # Implementation of basic git or git-related commands
-def is_executable(path: str) -> bool:
-    return os.access(path, os.X_OK)
-
-
-def find_executable(cli_ctxt: CommandLineContext, executable: str) -> Optional[str]:
-    base, ext = os.path.splitext(executable)
-
-    if (sys.platform == 'win32' or os.name == 'os2') and (ext != '.exe'):
-        executable = f"{executable}.exe"
-
-    if os.path.isfile(executable):
-        return executable
-
-    path = os.environ.get('PATH', os.defpath)
-    paths = path.split(os.pathsep)
-    for p in paths:
-        f = os.path.join(p, executable)
-        if os.path.isfile(f) and is_executable(f):
-            debug(cli_ctxt, f"find_executable({executable})", f"found {executable} at {f}")
-            return f
-    return None
+git_version = None
 
 
 def get_default_editor(cli_ctxt: CommandLineContext) -> Optional[str]:
@@ -1829,7 +1768,7 @@ def get_default_editor(cli_ctxt: CommandLineContext) -> Optional[str]:
             debug(cli_ctxt, "get_default_editor()", f"'{name}' is undefined")
         else:
             editor_repr = f"'{name}'{(' (' + editor + ')') if editor != name else ''}"
-            if not find_executable(cli_ctxt, editor):
+            if not utils.find_executable(cli_ctxt, editor):
                 debug(cli_ctxt, "get_default_editor()", f"{editor_repr} is not available")
                 if name == "$" + git_machete_editor_var:
                     # In this specific case, when GIT_MACHETE_EDITOR is defined but doesn't point to a valid executable,
@@ -1848,10 +1787,6 @@ def get_default_editor(cli_ctxt: CommandLineContext) -> Optional[str]:
 
     # This case is extremely unlikely on a modern Unix-like system.
     return None
-
-
-git_version = None
-
 
 def get_git_version(cli_ctxt: CommandLineContext) -> Tuple[int, int, int]:
     global git_version
@@ -1908,7 +1843,7 @@ def ensure_config_loaded(cli_ctxt: CommandLineContext) -> None:
     global config_cached
     if config_cached is None:
         config_cached = {}
-        for config_line in non_empty_lines(popen_git(cli_ctxt, "config", "--list")):
+        for config_line in utils.non_empty_lines(popen_git(cli_ctxt, "config", "--list")):
             k_v = config_line.split("=", 1)
             if len(k_v) == 2:
                 k, v = k_v
@@ -1939,7 +1874,7 @@ remotes_cached = None
 def remotes(cli_ctxt: CommandLineContext) -> List[str]:
     global remotes_cached
     if remotes_cached is None:
-        remotes_cached = non_empty_lines(popen_git(cli_ctxt, "remote"))
+        remotes_cached = utils.non_empty_lines(popen_git(cli_ctxt, "remote"))
     return remotes_cached
 
 
@@ -2248,7 +2183,7 @@ def contains_equivalent_tree(
 
     # `git log later_commit_sha ^earlier_commit_sha`
     # shows all commits reachable from later_commit_sha but NOT from earlier_commit_sha
-    intermediate_tree_shas = non_empty_lines(
+    intermediate_tree_shas = utils.non_empty_lines(
         popen_git(
             cli_ctxt,
             "log",
@@ -2263,14 +2198,14 @@ def contains_equivalent_tree(
     return result
 
 
-def create_branch(cli_ctxt: CommandLineContext, b: str, out_of_revision: str) -> None:
+def create_branch(cli_ctxt: CommandLineContext, b: str, out_of_revision: str) -> None:  # maybe move such a methods into github.py
     run_git(cli_ctxt, "checkout", "-b", b, out_of_revision)
     flush_caches()  # the repository state has changed b/c of a successful branch creation, let's defensively flush all the caches
 
 
 def log_shas(cli_ctxt: CommandLineContext, revision: str, max_count: Optional[int]) -> List[str]:
     opts = ([f"--max-count={str(max_count)}"] if max_count else []) + ["--format=%H", f"refs/heads/{revision}"]
-    return non_empty_lines(popen_git(cli_ctxt, "log", *opts))
+    return utils.non_empty_lines(popen_git(cli_ctxt, "log", *opts))
 
 
 MAX_COUNT_FOR_INITIAL_LOG = 10
@@ -2323,7 +2258,7 @@ def load_branches(cli_ctxt: CommandLineContext) -> None:
     tree_sha_by_commit_sha_cached = {}
 
     # Using 'committerdate:raw' instead of 'committerdate:unix' since the latter isn't supported by some older versions of git.
-    raw_remote = non_empty_lines(popen_git(cli_ctxt, "for-each-ref", "--format=%(refname)\t%(objectname)\t%(tree)\t%(committerdate:raw)", "refs/remotes"))
+    raw_remote = utils.non_empty_lines(popen_git(cli_ctxt, "for-each-ref", "--format=%(refname)\t%(objectname)\t%(tree)\t%(committerdate:raw)", "refs/remotes"))
     for line in raw_remote:
         values = line.split("\t")
         if len(values) != 4:
@@ -2335,7 +2270,7 @@ def load_branches(cli_ctxt: CommandLineContext) -> None:
         tree_sha_by_commit_sha_cached[commit_sha] = tree_sha
         committer_unix_timestamp_by_revision_cached[b] = int(committer_unix_timestamp_and_time_zone.split(' ')[0])
 
-    raw_local = non_empty_lines(popen_git(cli_ctxt, "for-each-ref", "--format=%(refname)\t%(objectname)\t%(tree)\t%(committerdate:raw)\t%(upstream)", "refs/heads"))
+    raw_local = utils.non_empty_lines(popen_git(cli_ctxt, "for-each-ref", "--format=%(refname)\t%(objectname)\t%(tree)\t%(committerdate:raw)\t%(upstream)", "refs/heads"))
 
     for line in raw_local:
         values = line.split("\t")
@@ -2367,7 +2302,7 @@ def get_sole_remote_branch(cli_ctxt: CommandLineContext, b: str) -> Optional[str
 def merged_local_branches(cli_ctxt: CommandLineContext) -> List[str]:
     return list(map(
         lambda b: re.sub("^refs/heads/", "", b),
-        non_empty_lines(popen_git(cli_ctxt, "for-each-ref", "--format=%(refname)", "--merged", "HEAD", "refs/heads"))
+        utils.non_empty_lines(popen_git(cli_ctxt, "for-each-ref", "--format=%(refname)", "--merged", "HEAD", "refs/heads"))
     ))
 
 
@@ -2383,7 +2318,7 @@ def get_hook_path(cli_ctxt: CommandLineContext, hook_name: str) -> str:
 def check_hook_executable(cli_ctxt: CommandLineContext, hook_path: str) -> bool:
     if not os.path.isfile(hook_path):
         return False
-    elif not is_executable(hook_path):
+    elif not utils.is_executable(hook_path):
         advice_ignored_hook = get_config_or_none(cli_ctxt, "advice.ignoredHook")
         if advice_ignored_hook != 'false':  # both empty and "true" is okay
             # The [33m color must be used to keep consistent with how git colors this advice for its built-in hooks.
@@ -2464,7 +2399,7 @@ def commits_between(cli_ctxt: CommandLineContext, earliest_exclusive: str, lates
     # Reverse the list, since `git log` by default returns the commits from the latest to earliest.
     return list(reversed(list(map(
         lambda x: tuple(x.split(":", 2)),  # type: ignore
-        non_empty_lines(popen_git(cli_ctxt, "log", "--format=%H:%h:%s", f"^{earliest_exclusive}", latest_inclusive, "--"))
+        utils.non_empty_lines(popen_git(cli_ctxt, "log", "--format=%H:%h:%s", f"^{earliest_exclusive}", latest_inclusive, "--"))
     ))))
 
 
@@ -2525,7 +2460,7 @@ def load_all_reflogs(cli_ctxt: CommandLineContext) -> None:
     all_branches = [f"refs/heads/{b}" for b in local_branches(cli_ctxt)] + \
                    [f"refs/remotes/{combined_counterpart_for_fetching_of_branch(cli_ctxt, b)}" for b in local_branches(cli_ctxt) if combined_counterpart_for_fetching_of_branch(cli_ctxt, b)]
     # The trailing '--' is necessary to avoid ambiguity in case there is a file called just exactly like one of the branches.
-    entries = non_empty_lines(popen_git(cli_ctxt, "reflog", "show", "--format=%gD\t%H\t%gs", *(all_branches + ["--"])))
+    entries = utils.non_empty_lines(popen_git(cli_ctxt, "reflog", "show", "--format=%gD\t%H\t%gs", *(all_branches + ["--"])))
     reflogs_cached = {}
     for entry in entries:
         values = entry.split("\t")
@@ -2570,7 +2505,7 @@ def get_latest_checkout_timestamps(cli_ctxt: CommandLineContext) -> Dict[str, in
     #   `--date=unix` is not available on some older versions of git)
     # %gs - reflog subject
     output = popen_git(cli_ctxt, "reflog", "show", "--format=%gd:%gs", "--date=raw")
-    for entry in non_empty_lines(output):
+    for entry in utils.non_empty_lines(output):
         pattern = "^HEAD@\\{([0-9]+) .+\\}:checkout: moving from (.+) to (.+)$"
         match = re.search(pattern, entry)
         if match:
@@ -2739,6 +2674,7 @@ def handle_untracked_branch(cli_ctxt: CommandLineContext, new_remote: str, b: st
 
 
 # Main
+
 alias_by_command: Dict[str, str] = {
     "diff": "d",
     "edit": "e",
@@ -2756,657 +2692,6 @@ command_groups: List[Tuple[str, List[str]]] = [
     ("Determine changes specific to the given branch", ["diff", "fork-point", "log"]),
     ("Update git history in accordance with the tree of branch dependencies", ["advance", "reapply", "slide-out", "squash", "traverse", "update"])
 ]
-
-short_docs: Dict[str, str] = {
-    "add": "Add a branch to the tree of branch dependencies",
-    "advance": "Fast-forward merge one of children to the current branch and then slide out this child",
-    "anno": "Manage custom annotations",
-    "delete-unmanaged": "Delete local branches that are not present in the definition file",
-    "diff": "Diff current working directory or a given branch against its computed fork point",
-    "discover": "Automatically discover tree of branch dependencies",
-    "edit": "Edit the definition file",
-    "file": "Display the location of the definition file",
-    "fork-point": "Display or override fork point for a branch",
-    "format": "Display docs for the format of the definition file",
-    "go": "Check out the branch relative to the position of the current branch, accepts down/first/last/next/root/prev/up argument",
-    "help": "Display this overview, or detailed help for a specified command",
-    "hooks": "Display docs for the extra hooks added by git machete",
-    "is-managed": "Check if the current branch is managed by git machete (mostly for scripts)",
-    "list": "List all branches that fall into one of pre-defined categories (mostly for internal use)",
-    "log": "Log the part of history specific to the given branch",
-    "reapply": "Rebase the current branch onto its computed fork point",
-    "show": "Show name(s) of the branch(es) relative to the position of a branch, accepts down/first/last/next/root/prev/up argument",
-    "slide-out": "Slide out the current branch and sync its downstream (child) branches with its upstream (parent) branch via rebase or merge",
-    "squash": "Squash the unique history of the current branch into a single commit",
-    "status": "Display formatted tree of branch dependencies, including info on their sync with upstream branch and with remote",
-    "traverse": "Walk through the tree of branch dependencies and rebase, merge, slide out, push and/or pull each branch one by one",
-    "update": "Sync the current branch with its upstream (parent) branch via rebase or merge",
-    "version": "Display the version and exit"
-}
-
-long_docs: Dict[str, str] = {
-    "add": """
-        <b>Usage: git machete add [-o|--onto=<target-upstream-branch>] [-R|--as-root] [-y|--yes] [<branch>]</b>
-
-        Adds the provided <branch> (or the current branch, if none specified) to the definition file.
-        If <branch> is provided but no local branch with the given name exists:
-        * if a remote branch of the same name exists in exactly one remote, then user is asked whether to check out this branch locally (as in `git checkout`),
-        * otherwise, user is asked whether it should be created as a new local branch.
-
-        If the definition file is empty or `-R`/`--as-root` is provided, the branch will be added as a root of the tree of branch dependencies.
-        Otherwise, the desired upstream (parent) branch can be specified with `-o`/`--onto`.
-        Neither of these options is mandatory, however; if both are skipped, git machete will try to automatically infer the target upstream.
-        If the upstream branch can be inferred, the user will be presented with inferred branch and asked to confirm.
-
-        Note: all the effects of `add` (except git branch creation) can as well be achieved by manually editing the definition file.
-
-        <b>Options:</b>
-          <b>-o, --onto=<target-upstream-branch></b>    Specifies the target parent branch to add the given branch onto.
-                                                 Cannot be specified together with `-R`/`--as-root`.
-
-          <b>-R, --as-root</b>                          Add the given branch as a new root (and not onto any other branch).
-                                                 Cannot be specified together with `-o`/`--onto`.
-
-          <b>-y, --yes</b>                              Don't ask for confirmation whether to create the branch or whether to add onto the inferred upstream.
-    """,
-    "advance": """
-        <b>Usage: git machete advance [-y|--yes]</b>
-
-        Fast forwards (as in `git merge --ff-only`) the current branch `C` to match its downstream `D`,
-        and subsequently slides out `D`. Both steps require manual confirmation unless `-y`/`--yes` is provided.
-
-        The downstream `C` is selected according to the following criteria:
-        * if `C` has exactly one downstream branch `d` whose tip is a descendant of `C`, and whose fork point is equal to `C` or is overridden
-          (basically: there's a green edge between `C` and `d`), then `d` is selected as `D`,
-        * if `C` has no downstream branches connected with a green edge to `C`, then `advance` fails,
-        * if `C` has more than one downstream branch connected with a green edge to `C`,
-          then user is asked to pick the branch to fast-forward merge into (similarly to what happens in `git machete go down`).
-          If `--yes` is specified, then `advance` fails.
-
-        As an example, if `git machete status --color=never --list-commits` is as follows:
-        <dim>
-          master
-          |
-          m-develop *
-            |
-            | Enable adding remote branch in the manner similar to git checkout
-            o-feature/add-from-remote
-            | |
-            | | Add support and sample for machete-post-slide-out hook
-            | o-feature/post-slide-out-hook
-            |
-            | Remove support for Python 2
-            | Remove support for Python 2 - 1st round of fixes
-            ?-chore/v3
-            |
-            | Apply Python2-compatible static typing
-            x-feature/types
-        </dim>
-        then running `git machete advance` will fast-forward the current branch `develop` to match `feature/add-from-remote`, and subsequently slide out the latter.
-        After `advance` completes, `status` will show:
-        <dim>
-          master
-          |
-          | Enable adding remote branch in the manner similar to git checkout
-          o-develop *
-            |
-            | Add support and sample for machete-post-slide-out hook
-            o-feature/post-slide-out-hook
-            |
-            | Remove support for Python 2
-            | Remove support for Python 2 - 1st round of fixes
-            ?-chore/v3
-            |
-            | Apply Python2-compatible static typing
-            x-feature/types
-        </dim>
-        Note that the current branch after the operation is still `develop`, just pointing to `feature/add-from-remote`'s tip now.
-
-        <b>Options:</b>
-          <b>-y, --yes</b>         Don't ask for confirmation whether to fast-forward the current branch or whether to slide-out the downstream.
-                            Fails if the current branch has more than one green-edge downstream branch.
-    """,
-    "anno": """
-        <b>Usage:
-          git machete anno [-b|--branch=<branch>] [<annotation text>]
-          git machete anno -H|--sync-github-prs</b>
-
-        If invoked without any argument, prints out the custom annotation for the given branch (or current branch, if none specified with `-b/--branch`).
-
-        If invoked with a single empty string argument, like:
-        <dim>$ git machete anno ''</dim>
-        then clears the annotation for the current branch (or a branch specified with `-b/--branch`).
-
-        If invoked with `-H` or `--sync-github-prs`, annotates the branches based on their corresponding GitHub PR numbers and authors.
-        Any existing annotations are overwritten for the branches that have an opened PR; annotations for the other branches remain untouched.
-
-        To allow GitHub API access for private repositories (and also to correctly identify the current user, even in case of public repositories),
-        a GitHub API token with `repo` scope is required, see `https://github.com/settings/tokens`. This will be resolved from the first of:
-        1. `GITHUB_TOKEN` env var,
-        2. current auth token from the `gh` GitHub CLI,
-        3. current auth token from the `hub` GitHub CLI.
-
-        In any other case, sets the annotation for the given/current branch to the given argument.
-        If multiple arguments are passed to the command, they are concatenated with a single space.
-
-        Note: all the effects of `anno` can be always achieved by manually editing the definition file.
-
-        <b>Options:</b>
-          <b>-b, --branch=<branch></b>      Branch to set the annotation for.
-          <b>-H, --sync-github-prs</b>      Annotate with GitHub PR numbers and authors where applicable.
-    """,
-    "delete-unmanaged": """
-        <b>Usage: git machete delete-unmanaged [-y|--yes]</b>
-
-        Goes one-by-one through all the local git branches that don't exist in the definition file,
-        and ask to delete each of them (with `git branch -d` or `git branch -D`) if confirmed by user.
-        No branch will be deleted unless explicitly confirmed by the user (or unless `-y/--yes` option is passed).
-
-        Note: this should be used with care since deleting local branches can sometimes make it impossible for `git machete` to properly figure out fork points.
-        See `git machete help fork-point` for more details.
-
-        <b>Options:</b>
-          <b>-y, --yes</b>          Don't ask for confirmation.
-    """,
-    "diff": """
-        <b>Usage: git machete d[iff] [-s|--stat] [<branch>]</b>
-
-        Runs `git diff` of the given branch tip against its fork point or, if none specified, of the current working tree against the fork point of the currently checked out branch.
-        See `git machete help fork-point` for more details on meaning of the "fork point".
-
-        Note: the branch in question does not need to occur in the definition file.
-
-        Options:
-          <b>-s, --stat</b>    Makes `git machete diff` pass `--stat` option to `git diff`, so that only summary (diffstat) is printed.
-    """,
-    "discover": """
-        <b>Usage: git machete discover [-C|--checked-out-since=<date>] [-l|--list-commits] [-r|--roots=<branch1>,<branch2>,...] [-y|--yes]</b>
-
-        Discovers and displays tree of branch dependencies using a heuristic based on reflogs and asks whether to overwrite the existing definition file with the new discovered tree.
-        If confirmed with a `y[es]` or `e[dit]` reply, backs up the current definition file (if it exists) as `$GIT_DIR/machete~` and saves the new tree under the usual `$GIT_DIR/machete` path.
-        If the reply was `e[dit]`, additionally an editor is opened (as in `git machete edit`) after saving the new definition file.
-
-        Options:
-          <b>-C, --checked-out-since=<date></b>   Only consider branches checked out at least once since the given date. <date> can be e.g. `2 weeks ago` or `2020-06-01`, as in `git log --since=<date>`.
-                                           If not present, the date is selected automatically so that around """ + str(MacheteClient.DISCOVER_DEFAULT_FRESH_BRANCH_COUNT) + """ branches are included.
-
-          <b>-l, --list-commits</b>               When printing the discovered tree, additionally lists the messages of commits introduced on each branch (as for `git machete status`).
-
-          <b>-r, --roots=<branch1,...></b>        Comma-separated list of branches that should be considered roots of trees of branch dependencies.
-                                           If not present, `master` is assumed to be a root.
-                                           Note that in the process of discovery, certain other branches can also be additionally deemed to be roots as well.
-
-          <b>-y, --yes</b>                        Don't ask for confirmation before saving the newly-discovered tree.
-                                           Mostly useful in scripts; not recommended for manual use.
-    """,
-    "edit": """
-        <b>Usage: git machete e[dit]</b>
-
-        Opens an editor and lets you edit the definition file manually.
-
-        The editor is determined by checking up the following locations:
-        * `$GIT_MACHETE_EDITOR`
-        * `$GIT_EDITOR`
-        * `$(git config core.editor)`
-        * `$VISUAL`
-        * `$EDITOR`
-        * `editor`
-        * `nano`
-        * `vi`
-        and selecting the first one that is defined and points to an executable file accessible on `PATH`.
-
-        Note that the above editor selection only applies for editing the definition file,
-        but not for any other actions that may be indirectly triggered by git-machete, including editing of rebase TODO list, commit messages etc.
-
-        The definition file can be always accessed and edited directly under path returned by `git machete file` (currently fixed to <git-directory>/machete).
-    """,
-    "file": """
-        <b>Usage: git machete file</b>
-
-        Outputs the absolute path of the machete definition file. Currently fixed to `<git-directory>/machete`.
-        Note: this won't always be just `<repo-root>/.git/machete` since e.g. submodules and worktrees have their git directories in different location.
-    """,
-    "fork-point": """
-        <b>Usage:
-          git machete fork-point [--inferred] [<branch>]
-          git machete fork-point --override-to=<revision>|--override-to-inferred|--override-to-parent [<branch>]
-          git machete fork-point --unset-override [<branch>]</b>
-
-        Note: in all three forms, if no <branch> is specified, the currently checked out branch is assumed.
-        The branch in question does not need to occur in the definition file.
-
-
-        Without any option, displays full SHA of the fork point commit for the <branch>.
-        Fork point of the given <branch> is the commit at which the history of the <branch> diverges from history of any other branch.
-
-        Fork point is assumed by many `git machete` commands as the place where the unique history of the <branch> starts.
-        The range of commits between the fork point and the tip of the given branch is, for instance:
-        * listed for each branch by `git machete status --list-commits`
-        * passed to `git rebase` by `git machete reapply`/`slide-out`/`traverse`/`update`
-        * provided to `git diff`/`log` by `git machete diff`/`log`.
-
-        `git machete` assumes fork point of <branch> is the most recent commit in the log of <branch> that has NOT been introduced on that very branch,
-        but instead occurs on a reflog (see help for `git reflog`) of some other, usually chronologically earlier, branch.
-        This yields a correct result in typical cases, but there are some situations
-        (esp. when some local branches have been deleted) where the fork point might not be determined correctly.
-        Thus, all rebase-involving operations (`reapply`, `slide-out`, `traverse` and `update`) run `git rebase` in the interactive mode,
-        unless told explicitly not to do so by `--no-interactive-rebase` flag, so that the suggested commit range can be inspected before the rebase commences.
-        Also, `reapply`, `slide-out`, `squash`, and `update` allow to specify the fork point explicitly by a command-line option.
-
-        `git machete fork-point` is different (and more powerful) than `git merge-base --fork-point`,
-        since the latter takes into account only the reflog of the one provided upstream branch,
-        while the former scans reflogs of all local branches and their remote tracking branches.
-        This makes git-machete's `fork-point` more resilient to modifications of the tree definition which change the upstreams of branches.
-
-
-        With `--override-to=<revision>`, sets up a fork point override for <branch>.
-        Fork point for <branch> will be overridden to the provided <revision> (commit) as long as the <branch> still points to (or is descendant of) the commit X
-        that <branch> pointed to at the moment the override is set up.
-        Even if revision is a symbolic name (e.g. other branch name or `HEAD~3`) and not explicit commit hash (like `a1b2c3ff`),
-        it's still resolved to a specific commit hash at the moment the override is set up (and not later when the override is actually used).
-        The override data is stored under `machete.overrideForkPoint.<branch>.to` and `machete.overrideForkPoint.<branch>.whileDescendantOf` git config keys.
-        Note: the provided fork point <revision> must be an ancestor of the current <branch> commit X.
-
-        With `--override-to-parent`, overrides fork point of the <branch> to the commit currently pointed by <branch>'s parent in the branch dependency tree.
-        Note: this will only work if <branch> has a parent at all (i.e. is not a root) and parent of <branch> is an ancestor of current <branch> commit X.
-
-        With `--inferred`, displays the commit that `git machete fork-point` infers to be the fork point of <branch>.
-        If there is NO fork point override for <branch>, this is identical to the output of `git machete fork-point`.
-        If there is a fork point override for <branch>, this is identical to the what the output of `git machete fork-point` would be if the override was NOT present.
-
-        With `--override-to-inferred` option, overrides fork point of the <branch> to the commit that `git machete fork-point` infers to be the fork point of <branch>.
-        Note: this piece of information is also displayed by `git machete status --list-commits` in case a yellow edge occurs.
-
-        With `--unset-override`, the fork point override for <branch> is unset.
-        This is simply done by removing the corresponding `machete.overrideForkPoint.<branch>.*` config entries.
-
-
-        <b>Note:</b> if an overridden fork point applies to a branch `B`, then it's considered to be <green>connected with a green edge</green> to its upstream (parent) `U`,
-        even if the overridden fork point of `B` is NOT equal to the commit pointed by `U`.
-    """,
-    "format": """
-        Note: there is no 'git machete format' command as such; 'format' is just a topic of 'git machete help'.
-
-        The format of the definition file should be as follows:
-        <dim>
-          develop
-              adjust-reads-prec PR #234
-                  block-cancel-order PR #235
-                      change-table
-                          drop-location-type
-              edit-margin-not-allowed
-                  full-load-gatling
-              grep-errors-script
-          master
-              hotfix/receipt-trigger PR #236
-        </dim>
-        In the above example `develop` and `master` are roots of the tree of branch dependencies.
-        Branches `adjust-reads-prec`, `edit-margin-not-allowed` and `grep-errors-script` are direct downstream branches for `develop`.
-        `block-cancel-order` is a downstream branch of `adjust-reads-prec`, `change-table` is a downstream branch of `block-cancel-order` and so on.
-
-        Every branch name can be followed (after a single space as a delimiter) by a custom annotation - a PR number in the above example.
-        The annotations don't influence the way `git machete` operates other than that they are displayed in the output of the `status` command.
-        Also see help for the `anno` command.
-
-        Tabs or any number of spaces can be used as indentation.
-        It's only important to be consistent wrt. the sequence of characters used for indentation between all lines.
-    """,
-    "go": """
-        <b>Usage: git machete g[o] <direction></b>
-        where <direction> is one of: `d[own]`, `f[irst]`, `l[ast]`, `n[ext]`, `p[rev]`, `r[oot]`, `u[p]`
-
-        Checks out the branch specified by the given direction relative to the currently checked out branch.
-        Roughly equivalent to `git checkout $(git machete show <direction>)`.
-        See `git machete help show` on more details on meaning of each direction.
-    """,
-    "help": """
-        <b>Usage: git machete help [<command>]</b>
-
-        Prints a summary of this tool, or a detailed info on a command if defined.
-    """,
-    "hooks": """
-        As with the standard git hooks, git-machete looks for its own specific hooks in `$GIT_DIR/hooks/*` (or `$(git config core.hooksPath)/*`, if set).
-
-        Note: `hooks` is not a command as such, just a help topic (there is no `git machete hooks` command).
-
-        * <b>machete-post-slide-out <new-upstream> <lowest-slid-out-branch> [<new-downstreams>...]</b>
-            The hook that is executed after a branch (or possibly multiple branches, in case of `slide-out`)
-            is slid out by `advance`, `slide-out` or `traverse`.
-
-            At least two parameters (branch names) are passed to the hook:
-            * <b><new-upstream></b> is the upstream of the branch that has been slid out,
-              or in case of multiple branches being slid out - the upstream of the highest slid out branch;
-            * <b><lowest-slid-out-branch></b> is the branch that has been slid out,
-              or in case of multiple branches being slid out - the lowest slid out branch;
-            * <b><new-downstreams></b> are all the following (possibly zero) parameters,
-              which correspond to all original downstreams of <lowest-slid-out-branch>, now reattached as the downstreams of <new-upstream>.
-              Note that this may be zero, one, or multiple branches.
-
-            Note: the hook, if present, is executed:
-            * zero or once during a `advance` execution (depending on whether the slide-out has been confirmed or not),
-            * exactly once during a `slide-out` execution (even if multiple branches are slid out),
-            * zero or more times during `traverse` (every time a slide-out operation is confirmed).
-
-            If the hook returns a non-zero exit code, then the execution of the command is aborted,
-            i.e. `slide-out` won't attempt rebase of the new downstream branches and `traverse` won't continue the traversal.
-            In case of `advance` there is no difference (other than exit code of the entire `advance` command being non-zero),
-            since slide-out is the last operation that happens within `advance`.
-            Note that non-zero exit code of the hook doesn't cancel the effects of slide-out itself, only the subsequent operations.
-            The hook is executed only once the slide-out is complete and can in fact rely on .git/machete file being updated to the new branch layout.
-
-        * <b>machete-pre-rebase <new-base> <fork-point-hash> <branch-being-rebased></b>
-            The hook that is executed before rebase is run during `reapply`, `slide-out`, `traverse` and `update`.
-            Note that it is NOT executed by `squash` (despite its similarity to `reapply`), since no rebase is involved in `squash`.
-
-            The parameters are exactly the three revisions that are passed to `git rebase --onto`:
-            1. what is going to be the new base for the rebased commits,
-            2. what is the fork point - the place where the rebased history diverges from the upstream history,
-            3. what branch is rebased.
-            If the hook returns a non-zero exit code, the entire rebase is aborted.
-
-            Note: this hook is independent from git's standard `pre-rebase` hook.
-            If machete-pre-rebase returns zero, the execution flow continues to `git rebase`, which may also run `pre-rebase` hook if present.
-            `machete-pre-rebase` is thus always launched before `pre-rebase`.
-
-        * <b>machete-status-branch <branch-name></b>
-            The hook that is executed for each branch displayed during `discover`, `status` and `traverse`.
-
-            The standard output of this hook is displayed at the end of the line, after branch name, (optionally) custom annotation and (optionally) remote sync-ness status.
-            Standard error is ignored. If the hook returns a non-zero exit code, both stdout and stderr are ignored, and printing the status continues as usual.
-
-            Note: the hook is always invoked with `ASCII_ONLY` variable passed into the environment.
-            If `status` runs in ASCII-only mode (i.e. if `--color=auto` and stdout is not a terminal, or if `--color=never`), then `ASCII_ONLY=true`, otherwise `ASCII_ONLY=false`.
-
-        Please see hook_samples/ directory of git-machete project for examples.
-        An example of using the standard git `post-commit` hook to `git machete add` branches automatically is also included.
-    """,
-    "is-managed": """
-        <b>Usage: git machete is-managed [<branch>]</b>
-
-        Returns with zero exit code if the given branch (or current branch, if none specified) is managed by git-machete (i.e. listed in .git/machete).
-
-        Returns with a non-zero exit code in case:
-        * the <branch> is provided but isn't managed, or
-        * the <branch> isn't provided and the current branch isn't managed, or
-        * the <branch> isn't provided and there's no current branch (detached HEAD).
-    """,
-    "list": """
-        <b>Usage: git machete list <category></b>
-        where <category> is one of: `addable`, `managed`, `slidable`, `slidable-after <branch>`, `unmanaged`, `with-overridden-fork-point`.
-
-        Lists all branches that fall into one of the specified categories:
-        * `addable`: all branches (local or remote) than can be added to the definition file,
-        * `managed`: all branches that appear in the definition file,
-        * `slidable`: all managed branches that have an upstream and can be slid out with `slide-out` command
-        * `slidable-after <branch>`: the downstream branch of the <branch>, if it exists and is the only downstream of <branch> (i.e. the one that can be slid out immediately following <branch>),
-        * `unmanaged`: all local branches that don't appear in the definition file,
-        * `with-overridden-fork-point`: all local branches that have a fork point override set up (even if this override does not affect the location of their fork point anymore).
-
-        This command is generally not meant for a day-to-day use, it's mostly needed for the sake of branch name completion in shell.
-    """,
-    "log": """
-        <b>Usage: git machete l[og] [<branch>]</b>
-
-        Runs `git log` for the range of commits from tip of the given branch (or current branch, if none specified) back to its fork point.
-        See `git machete help fork-point` for more details on meaning of the "fork point".
-
-        Note: the branch in question does not need to occur in the definition file.
-    """,
-    "reapply": """
-        <b>Usage: git machete reapply [-f|--fork-point=<fork-point-commit>]</b>
-
-        Interactively rebase the current branch on the top of its computed fork point.
-        The chunk of the history to be rebased starts at the automatically computed fork point of the current branch by default, but can also be set explicitly by `--fork-point`.
-        See `git machete help fork-point` for more details on meaning of the "fork point".
-
-        Note: the current reapplied branch does not need to occur in the definition file.
-
-        Tip: `reapply` can be used for squashing the commits on the current branch to make history more condensed before push to the remote,
-        but there is also dedicated `squash` command that achieves the same goal without running `git rebase`.
-
-        <b>Options:</b>
-          <b>-f, --fork-point=<fork-point-commit></b>    Specifies the alternative fork point commit after which the rebased part of history is meant to start.
-    """,
-    "show": """
-        <b>Usage: git machete show <direction> [<branch>]</b>
-        where <direction> is one of: `c[urrent]`, `d[own]`, `f[irst]`, `l[ast]`, `n[ext]`, `p[rev]`, `r[oot]`, `u[p]`
-        displayed relative to target <branch>, or the current checked out branch if <branch> is unspecified.
-
-        Outputs name of the branch (or possibly multiple branches, in case of `down`) that is:
-
-        * `current`: the current branch; exits with a non-zero status if none (detached HEAD)
-        * `down`:    the direct children/downstream branch of the target branch.
-        * `first`:   the first downstream of the root branch of the target branch (like `root` followed by `next`), or the root branch itself if the root has no downstream branches.
-        * `last`:    the last branch in the definition file that has the same root as the target branch; can be the root branch itself if the root has no downstream branches.
-        * `next`:    the direct successor of the target branch in the definition file.
-        * `prev`:    the direct predecessor of the target branch in the definition file.
-        * `root`:    the root of the tree where the target branch is located. Note: this will typically be something like `develop` or `master`, since all branches are usually meant to be ultimately merged to one of those.
-        * `up`:      the direct parent/upstream branch of the target branch.
-    """,
-    "slide-out": """
-        <b>Usage: git machete slide-out [-d|--down-fork-point=<down-fork-point-commit>] [-M|--merge] [-n|--no-edit-merge|--no-interactive-rebase] <branch> [<branch> [<branch> ...]]</b>
-
-        Removes the given branch (or multiple branches) from the branch tree definition.
-        Then synchronizes the downstream (child) branches of the last specified branch on the top of the upstream (parent) branch of the first specified branch.
-        Sync is performed either by rebase (default) or by merge (if `--merge` option passed).
-
-        The most common use is to slide out a single branch whose upstream was a `develop`/`master` branch and that has been recently merged.
-
-        Since this tool is designed to perform only one single rebase/merge at the end, provided branches must form a chain, i.e. all of the following conditions must be met:
-        * for i=1..N-1, (i+1)-th branch must be the only downstream (child) branch of the i-th branch,
-        * all provided branches must have an upstream branch (so, in other words, roots of branch dependency tree cannot be slid out).
-
-        For example, let's assume the following dependency tree:
-        <dim>
-          develop
-              adjust-reads-prec
-                  block-cancel-order
-                      change-table
-                          drop-location-type
-                      add-notification
-        </dim>
-        And now let's assume that `adjust-reads-prec` and later `block-cancel-order` were merged to develop.
-        After running `git machete slide-out adjust-reads-prec block-cancel-order` the tree will be reduced to:
-        <dim>
-          develop
-              change-table
-                  drop-location-type
-              add-notification
-        </dim>
-        and `change-table` and `add-notification` will be rebased onto develop (fork point for this rebase is configurable, see `-d` option below).
-
-        Note: This command doesn't delete any branches from git, just removes them from the tree of branch dependencies.
-
-        <b>Options:</b>
-          <b>-d, --down-fork-point=<down-fork-point-commit></b>    If updating by rebase, specifies the alternative fork point for downstream branches for the operation.
-                                                            `git machete fork-point` overrides for downstream branches are recommended over use of this option.
-                                                            See also doc for `--fork-point` option in `git machete help reapply` and `git machete help update`.
-                                                            Not allowed if updating by merge.
-
-          <b>-M, --merge</b>                                       Update the downstream branch by merge rather than by rebase.
-
-          <b>-n</b>                                                If updating by rebase, equivalent to `--no-interactive-rebase`. If updating by merge, equivalent to `--no-edit-merge`.
-
-          <b>--no-edit-merge</b>                                   If updating by merge, skip opening the editor for merge commit message while doing `git merge` (i.e. pass `--no-edit` flag to underlying `git merge`).
-                                                            Not allowed if updating by rebase.
-
-          <b>--no-interactive-rebase</b>                           If updating by rebase, run `git rebase` in non-interactive mode (without `-i/--interactive` flag).
-                                                            Not allowed if updating by merge.
-    """,
-    "squash": """
-        <b>Usage: git machete squash [-f|--fork-point=<fork-point-commit>]</b>
-
-        Squashes the commits belonging uniquely to the current branch into a single commit.
-        The chunk of the history to be squashed starts at the automatically computed fork point of the current branch by default, but can also be set explicitly by `--fork-point`.
-        See `git machete help fork-point` for more details on meaning of the "fork point".
-        The message for the squashed is taken from the earliest squashed commit, i.e. the commit directly following the fork point.
-
-        Note: the current reapplied branch does not need to occur in the definition file.
-
-        Tip: for more complex scenarios that require rewriting the history of current branch, see `reapply` and `update`.
-
-        <b>Options:</b>
-          <b>-f, --fork-point=<fork-point-commit></b>    Specifies the alternative fork point commit after which the squashed part of history is meant to start.
-    """,
-    "status": """
-        <b>Usage: git machete s[tatus] [--color=WHEN] [-l|--list-commits] [-L|--list-commits-with-hashes] [--no-detect-squash-merges]</b>
-
-        Displays a tree-shaped status of the branches listed in the definition file.
-
-        Apart from simply ASCII-formatting the definition file, this also:
-
-        * colors the edges between upstream (parent) and downstream (children) branches:
-
-          - <b><red>red edge</red></b> means that the downstream branch tip is <b>not a direct descendant</b> of the upstream branch tip,
-
-          - <b><yellow>yellow edge</yellow></b> means that the downstream branch tip is a <b>direct descendant</b> of the upstream branch tip,
-            but the fork point (see help on `fork-point`) of the downstream branch is <b>not equal</b> to the upstream branch tip,
-
-          - <b><green>green edge</green></b> means that the downstream branch tip is a <b>direct descendant</b> of the upstream branch tip
-            and the fork point of the downstream branch is <b>equal</b> to the upstream branch tip,
-
-          - <b><dim>grey/dimmed edge</dim></b> means that the downstream branch has been <b>merged</b> to the upstream branch,
-            detected by commit equivalency (default), or by strict detection of merge commits (if `--no-detect-squash-merges` passed).
-
-
-        * prints `(untracked/ahead of <remote>/behind <remote>/diverged from [& older than] <remote>)` message if the branch is not in sync with its remote counterpart;
-
-        * displays the custom annotations (see help on `format` and `anno`) next to each branch, if present;
-
-        * displays the output of `machete-status-branch` hook (see help on `hooks`), if present;
-
-        * optionally lists commits introduced on each branch if `-l`/`--list-commits` or `-L`/`--list-commits-with-hashes` is supplied.
-
-        The currently checked-out branch is underlined.
-
-        In case of yellow edge, use `-l` flag to show the exact location of the inferred fork point
-        (which indicates e.g. what range of commits is going to be rebased when the branch is updated).
-        The inferred fork point can be always overridden manually, see help on `fork-point`.
-
-        Grey/dimmed edge suggests that the downstream branch can be slid out (see help on `slide-out` and `traverse`).
-
-        Using colors can be disabled with a `--color` flag set to `never`.
-        With `--color=always`, git machete always emits colors and with `--color=auto`, it emits colors only when standard output is connected to a terminal.
-        `--color=auto` is the default. When colors are disabled, relation between branches is represented in the following way:
-        <dim>
-          <branch0>
-          |
-          o-<branch1> # green (in sync with parent)
-          | |
-          | x-<branch2> # red (not in sync with parent)
-          |   |
-          |   ?-<branch3> # yellow (in sync with parent, but parent is not the fork point)
-          |
-          m-<branch4> # grey (merged to parent)
-        </dim>
-        <b>Options:</b>
-          <b>--color=WHEN</b>                      Colorize the output; WHEN can be `always`, `auto` (default; i.e. only if stdout is a terminal), or `never`.
-
-          <b>-l, --list-commits</b>                Additionally list the commits introduced on each branch.
-
-          <b>-L, --list-commits-with-hashes</b>    Additionally list the short hashes and messages of commits introduced on each branch.
-
-          <b>--no-detect-squash-merges</b>         Only consider "strict" (fast-forward or 2-parent) merges, rather than rebase/squash merges,
-                                            when detecting if a branch is merged into its upstream (parent).
-    """,
-    "traverse": """
-        <b>Usage: git machete traverse [-F|--fetch] [-l|--list-commits] [-M|--merge]
-                                       [-n|--no-edit-merge|--no-interactive-rebase] [--no-detect-squash-merges]
-                                       [--[no-]push] [--[no-]push-untracked]
-                                       [--return-to=WHERE] [--start-from=WHERE] [-w|--whole] [-W] [-y|--yes]</b>
-
-        Traverses the branch dependency tree in pre-order (i.e. simply in the order as they occur in the definition file) and for each branch:
-        * detects if the branch is merged to its parent/upstream
-          - by commit equivalency (default), or by strict detection of merge commits (if `--no-detect-squash-merges` passed),
-          - if so, asks the user whether to slide out the branch from the dependency tree (typically branches are longer needed after they're merged);
-        * otherwise, if the branch is not in "green" sync with its parent/upstream (see help for `status`):
-          - asks the user whether to rebase (default) or merge (if `--merge` passed) the branch onto into its upstream branch - equivalent to `git machete update` with no `--fork-point` option passed;
-
-        * if the branch is not tracked on a remote, is ahead of its remote counterpart, or diverged from the counterpart & has newer head commit than the counterpart:
-          - asks the user whether to push the branch (possibly with `--force-with-lease` if the branches diverged);
-        * otherwise, if the branch diverged from the remote counterpart & has older head commit than the counterpart:
-          - asks the user whether to `git reset --keep` the branch to its remote counterpart
-        * otherwise, if the branch is behind its remote counterpart:
-          - asks the user whether to pull the branch;
-
-        * and finally, if any of the above operations has been successfully completed:
-          - prints the updated `status`.
-
-        Note that even if the traverse flow is stopped (typically due to merge/rebase conflicts), running `git machete traverse` after the merge/rebase is finished will pick up the walk where it stopped.
-        In other words, there is no need to explicitly ask to "continue" as it is the case with e.g. `git rebase`.
-
-        <b>Options:</b>
-          <b>-F, --fetch</b>                  Fetch the remotes of all managed branches at the beginning of traversal (no `git pull` involved, only `git fetch`).
-
-          <b>-l, --list-commits</b>           When printing the status, additionally list the messages of commits introduced on each branch.
-
-          <b>-M, --merge</b>                  Update by merge rather than by rebase.
-
-          <b>-n</b>                           If updating by rebase, equivalent to `--no-interactive-rebase`. If updating by merge, equivalent to `--no-edit-merge`.
-
-          <b>--no-detect-squash-merges</b>    Only consider "strict" (fast-forward or 2-parent) merges, rather than rebase/squash merges,
-                                       when detecting if a branch is merged into its upstream (parent).
-
-          <b>--no-edit-merge</b>              If updating by merge, skip opening the editor for merge commit message while doing `git merge` (i.e. pass `--no-edit` flag to underlying `git merge`).
-                                       Not allowed if updating by rebase.
-
-          <b>--no-interactive-rebase</b>      If updating by rebase, run `git rebase` in non-interactive mode (without `-i/--interactive` flag).
-                                       Not allowed if updating by merge.
-
-          <b>--no-push</b>                    Do not push any (neither tracked nor untracked) branches to remote, re-enable via `--push`.
-
-          <b>--no-push-untracked</b>          Do not push untracked branches to remote, re-enable via `--push-untracked`.
-
-          <b>--push</b>                       Push all (both tracked and untracked) branches to remote - default behavior.
-
-          <b>--push-untracked</b>             Push untracked branches to remote - default behavior.
-
-          <b>--return-to=WHERE</b>            Specifies the branch to return after traversal is successfully completed; WHERE can be `here` (the current branch at the moment when traversal starts),
-                                       `nearest-remaining` (nearest remaining branch in case the `here` branch has been slid out by the traversal)
-                                       or `stay` (the default - just stay wherever the traversal stops).
-                                       Note: when user quits by `q/yq` or when traversal is stopped because one of git actions fails, the behavior is always `stay`.
-
-          <b>--start-from=WHERE</b>           Specifies the branch to start the traversal from; WHERE can be `here` (the default - current branch, must be managed by git-machete),
-                                       `root` (root branch of the current branch, as in `git machete show root`) or `first-root` (first listed managed branch).
-
-          <b>-w, --whole</b>                  Equivalent to `-n --start-from=first-root --return-to=nearest-remaining`;
-                                       useful for quickly traversing & syncing all branches (rather than doing more fine-grained operations on the local section of the branch tree).
-
-          <b>-W</b>                           Equivalent to `--fetch --whole`; useful for even more automated traversal of all branches.
-
-          <b>-y, --yes</b>                    Don't ask for any interactive input, including confirmation of rebase/push/pull. Implies `-n`.
-    """,
-    "update": """
-        <b>Usage: git machete update [-f|--fork-point=<fork-point-commit>] [-M|--merge] [-n|--no-edit-merge|--no-interactive-rebase]</b>
-
-        Synchronizes the current branch with its upstream (parent) branch either by rebase (default) or by merge (if `--merge` option passed).
-
-        If updating by rebase, interactively rebases the current branch on the top of its upstream (parent) branch.
-        The chunk of the history to be rebased starts at the fork point of the current branch, which by default is inferred automatically, but can also be set explicitly by `--fork-point`.
-        See `git machete help fork-point` for more details on meaning of the "fork point".
-
-        If updating by merge, merges the upstream (parent) branch into the current branch.
-
-        <b>Options:</b>
-          <b>-f, --fork-point=<fork-point-commit></b>    If updating by rebase, specifies the alternative fork point commit after which the rebased part of history is meant to start.
-                                                  Not allowed if updating by merge.
-
-          <b>-M, --merge</b>                             Update by merge rather than by rebase.
-
-          <b>-n</b>                                      If updating by rebase, equivalent to `--no-interactive-rebase`. If updating by merge, equivalent to `--no-edit-merge`.
-
-          <b>--no-edit-merge</b>                         If updating by merge, skip opening the editor for merge commit message while doing `git merge` (i.e. pass `--no-edit` flag to underlying `git merge`).
-                                                  Not allowed if updating by rebase.
-
-          <b>--no-interactive-rebase</b>                 If updating by rebase, run `git rebase` in non-interactive mode (without `-i/--interactive` flag).
-                                                  Not allowed if updating by merge.
-    """,
-    "version": """
-        <b>Usage: git machete version</b>
-
-        Prints the version and exits.
-    """
-}
 
 
 def usage(c: str = None) -> None:
@@ -3454,6 +2739,7 @@ def launch(orig_args: List[str]) -> None:
 
     cli_ctxt = CommandLineContext()
     machete_client = MacheteClient(cli_ctxt)
+
 
     if sys.version_info.major == 2 or (sys.version_info.major == 3 and sys.version_info.minor < 6):
         version_str = f"{sys.version_info.major}.{sys.version_info.minor}.{sys.version_info.micro}"
@@ -3733,9 +3019,9 @@ def launch(orig_args: List[str]) -> None:
                 def strip_first_fragment(rb: str) -> str:
                     return re.sub("^[^/]+/", "", rb)
 
-                remote_counterparts_of_local_branches = map_truthy_only(lambda b: combined_counterpart_for_fetching_of_branch(cli_ctxt, b), local_branches(cli_ctxt))
-                qualifying_remote_branches = excluding(remote_branches(cli_ctxt), remote_counterparts_of_local_branches)
-                res = excluding(local_branches(cli_ctxt), machete_client.managed_branches) + list(map(strip_first_fragment, qualifying_remote_branches))
+                remote_counterparts_of_local_branches = utils.map_truthy_only(lambda b: combined_counterpart_for_fetching_of_branch(cli_ctxt, b), local_branches(cli_ctxt))
+                qualifying_remote_branches = utils.excluding(remote_branches(cli_ctxt), remote_counterparts_of_local_branches)
+                res = utils.excluding(local_branches(cli_ctxt), machete_client.managed_branches) + list(map(strip_first_fragment, qualifying_remote_branches))
             elif param == "managed":
                 res = machete_client.managed_branches
             elif param == "slidable":
@@ -3745,7 +3031,7 @@ def launch(orig_args: List[str]) -> None:
                 machete_client.expect_in_managed_branches(b_arg)
                 res = machete_client.slidable_after(b_arg)
             elif param == "unmanaged":
-                res = excluding(local_branches(cli_ctxt), machete_client.managed_branches)
+                res = utils.excluding(local_branches(cli_ctxt), machete_client.managed_branches)
             elif param == "with-overridden-fork-point":
                 res = list(filter(lambda b: machete_client.has_any_fork_point_override_config(b), local_branches(cli_ctxt)))
 
@@ -3825,9 +3111,9 @@ def launch(orig_args: List[str]) -> None:
     except StopTraversal:
         pass
     finally:
-        if initial_current_directory and not directory_exists(initial_current_directory):
+        if initial_current_directory and not utils.directory_exists(initial_current_directory):
             nearest_existing_parent_directory = initial_current_directory
-            while not directory_exists(nearest_existing_parent_directory):
+            while not utils.directory_exists(nearest_existing_parent_directory):
                 nearest_existing_parent_directory = os.path.join(nearest_existing_parent_directory, os.path.pardir)
             warn(f"current directory {initial_current_directory} no longer exists, "
                  f"the nearest existing parent directory is {os.path.abspath(nearest_existing_parent_directory)}")
