+++ git_machete/utils.py
@@ -1,0 +1,54 @@
+
+from typing import Any, Callable, Iterable, List, Optional, Tuple, TypeVar
+
+import os
+
+T = TypeVar('T')
+
+
+def excluding(iterable: Iterable[T], s: Iterable[T]) -> List[T]:
+    return list(filter(lambda x: x not in s, iterable))
+
+
+def flat_map(func: Callable[[T], List[T]], iterable: Iterable[T]) -> List[T]:
+    return sum(map(func, iterable), [])
+
+
+def map_truthy_only(func: Callable[[T], Optional[T]], iterable: Iterable[T]) -> List[T]:
+    return list(filter(None, map(func, iterable)))
+
+
+def non_empty_lines(s: str) -> List[str]:
+    return list(filter(None, s.split("\n")))
+
+
+# Converts a lambda accepting N arguments to a lambda accepting one argument, an N-element tuple.
+# Name matching Scala's `tupled` on `FunctionX`.
+def tupled(f: Callable[..., T]) -> Callable[[Any], T]:
+    return lambda tple: f(*tple)
+
+
+def get_second(pair: Tuple[str, str]) -> str:
+    a, b = pair
+    return b
+
+
+def directory_exists(path: str) -> bool:
+    try:
+        # Note that os.path.isdir itself (without os.path.abspath) isn't reliable
+        # since it returns a false positive (True) for the current directory when if it doesn't exist
+        return os.path.isdir(os.path.abspath(path))
+    except OSError:
+        return False
+
+
+def current_directory_or_none() -> Optional[str]:
+    try:
+        return os.getcwd()
+    except OSError:
+        # This happens when current directory does not exist (typically: has been deleted)
+        return None
+
+
+def is_executable(path: str) -> bool:
+    return os.access(path, os.X_OK)
